requirements.txt
wkhtmltopdf 0.12.6
pdfkit-1.0.0